# cmd/gateway/openapi.yaml
openapi: 3.0.3
info:
  title: AI Gateway
  version: 1.7.0
  description: |
    Unified AI Gateway — OpenAI / Anthropic / Responses under one domain.
    Gateway auth: Bearer sk-gw-* or admin session JWT on /v1/*.
    Admin: Bearer JWT on /api/*. Errors: {error:{message,type}}.
servers:
  - url: http://localhost:8080
    description: local
  - url: https://ai.example.com
    description: production
paths:
...
  /api/admin/users:
    get:
      summary: List dashboard users
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: users }
    post:
      summary: Create dashboard user (admin only, audited)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string }
                password: { type: string }
                role: { type: string, enum: [admin, support, member, readonly] }
                display_name: { type: string }
      responses:
        '201': { description: created }
        '400': { description: invalid input or role }
        '403': { description: forbidden (admin only) }
  /api/admin/users/me:
    get:
      summary: Current authenticated dashboard user (id/me resolves before {id})
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: me }
  /api/admin/users/{id}:
    put:
      summary: Update a dashboard user (role/display_name/password/disabled; admin only)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [admin, support, member, readonly] }
                display_name: { type: string }
                password: { type: string }
                disabled: { type: boolean }
      responses:
        '200': { description: updated }
        '400': { description: invalid role }
        '403': { description: forbidden }
    delete:
      summary: Delete a dashboard user (admin only)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: deleted }
  /api/admin/users/{id}/reset-password:
    post:
      summary: Reset a dashboard user's password (admin only; revokes their sessions)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        '200': { description: ok }
        '400': { description: password required }
        '403': { description: forbidden }
...
  /api/lb/rules:
    get:
      summary: List load-balancer routing rules (admin)
      description: |
        Each rule maps a model name to an ORDERED provider group. Rotation
        happens across requests (round-robin by position). There is no
        automatic failover: members never chain within one request. Qualified
        model IDs ("provider/model") and X-Provider headers bypass rules.
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Rules with ordered members (joined provider metadata)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LBRule' }
  /api/lb/rules/{model}:
    put:
      summary: Create or replace the provider group for a model (admin)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: model
          in: path
          required: true
          schema: { type: string, maxLength: 256, example: gpt-4o-mini }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providers]
              properties:
                providers:
                  type: array
                  maxItems: 50
                  minItems: 1
                  items: { type: string }
                  description: Provider IDs in round-robin order.
      responses:
        '201': { $ref: '#/components/responses/LBRuleSaved' }
        '400': { description: Unknown/duplicate provider or invalid body }
        '403': { description: Org-scoped caller referencing another org's provider }
    delete:
      summary: Delete the routing rule for a model (admin)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: model
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
...